refactor: update directory naming
