+++ lib/screens/main_screen.dart
@@ -3,7 +3,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_training/entity/weather_condition.dart';
 import 'package:flutter_training/extension/yumemi_weather_error_extension.dart';
-import 'package:flutter_training/state/weather_forecast_notifier.dart';
+import 'package:flutter_training/notifier/weather_forecast_notifier.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
 import 'package:yumemi_weather/yumemi_weather.dart';
